slides 13-16

+++ C-programming/slide14.c
@@ -2,11 +2,11 @@
 
 int main()
 {
-    int this_is_a_number;
+    float this_is_a_number;
 
     printf("Please enter a number: ");
-    scanf("%d", &this_is_a_number);
-    printf("You entered %d\n", this_is_a_number);
+    scanf("%f", &this_is_a_number);
+    printf("You entered %f\n", this_is_a_number);
     getchar();
     return 0;
 }
